refactor(mcp-meta): move ConfigSchemaMeta to meta/rjsf.py
Decouples RJSF schema generation from tool-specific config helpers.
